Bug #33117410 ITEM_SUM_BIT::ADD_BITS: ASSERTION `(I * 8) + BIT < M_DIGIT_CNT_CARD' FAILED. AND HEAP-BUFFER-OVERFLOW [noclose]
Bit functions in window expressions will assert that the runtime size
of a bit mask is not bigger than the resolve time size. We found
several violations of this:

The function encrypt() could get an overflow in the resolve phase when
computing the maximum size of the result.  Change my_aes_get_size() to
return longlong rather than int.

The functions convert(), concat(), concat_ws(), export_set(),
insert(), replace() and weight_string() did not compute max result
length properly for the binary character set.

During resolving of REPLACE(str, from_str, to_str) we assumed that the
max_length of 'from_str' would be replaced for each match in 'str'.
This was not correct, the *actual length* of from_str may be only one,
so during evaluation we may replace the entire 'str' with with
multiple copies of 'to_str'.

The function compress() had an ad hoc computation of max result
length. Use compressBound from the zlib library instead.

Boolean expressions have max_length == 1. For this to work in all
contexts we need to set unsigned_flag = true. We also need to use
MYSQL_TYPE_LONGLONG rather than MYSQL_TYPE_LONG as data type in order
to have mysqlxtest interpret results properly.

The function my_decimal_precision_to_length_no_truncation() is used to
set 'max_length' for many numeric operations. Ensure that we never
return zero.

The function Item_hex_string::val_int() does not signal an error in
case of bad input, it only signals a warning. The return value of -1
for this case could break assumptions done during the resolve phase
about length/precision for float/decimal values. So return zero
instead.

Item_func_mod::result_precision() used max_length of its arguments for
calculating its own max_length. That does not work for binary inputs.

+++ mysys/my_aes_openssl.cc
@@ -204,7 +204,7 @@ aes_error:
   return MY_AES_BAD_DATA;
 }
 
-int my_aes_get_size(uint32 source_length, my_aes_opmode opmode) {
+longlong my_aes_get_size(uint32 source_length, my_aes_opmode opmode) {
   const EVP_CIPHER *cipher = aes_evp_type(opmode);
   size_t block_size;
